MCR-1207 move code to MCRURIResolver as internal class
fixed codacy warning
